use carousel component for related products
