+++ client/src/components/relatedItems/RelatedProducts.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ProductCard from './ProductCard.jsx';
+import Carousel from './Carousel.jsx';
 import ComparisonModal from './ComparisonModal.jsx';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faStar, faChevronLeft, faChevronRight } from '@fortawesome/free-solid-svg-icons'
@@ -16,9 +16,6 @@ class RelatedProducts extends React.Component {
 
     this.getRelatedProducts = this.getRelatedProducts.bind(this);
     this.handleActionButton = this.handleActionButton.bind(this);
-    this.slideLeft = this.slideLeft.bind(this);
-    this.slideRight = this.slideRight.bind(this);
-
 
   }
 
@@ -41,17 +38,7 @@ class RelatedProducts extends React.Component {
     this.setState({showModal: newState, comparisonProductId: productId});
   }
 
-  slideLeft() {
-    var scroller = document.getElementById('carousel-content');
-    scroller.scrollLeft = scroller.scrollLeft - 260;
-    console.log(scroller.scrollLeft);
-  }
 
-  slideRight() {
-    var scroller = document.getElementById('carousel-content');
-    scroller.scrollLeft = scroller.scrollLeft + 260;
-    console.log(scroller.scrollLeft);
-  }
 
   componentDidMount() {
     this.getRelatedProducts(this.props.productId);
@@ -66,14 +53,7 @@ class RelatedProducts extends React.Component {
     return (
     <div id='related-products'>
       <h3>Related Products</h3>
-      <div id='carousel'>
-        <FontAwesomeIcon size='2x' icon={faChevronLeft} className='carousel-button left' onClick={this.slideLeft}/>
-        <div id='carousel-content'>
-          {this.state.relatedProducts.map((productId) =>
-            <ProductCard key={productId.toString()} productId={productId} setNewProductId={this.props.setNewProductId} actionButtonIcon={faStar} actionClick={this.handleActionButton}/>)}
-        </div>
-        <FontAwesomeIcon size='2x' icon={faChevronRight} className='carousel-button right' onClick={this.slideRight}/>
-      </div>
+      <Carousel products={this.state.relatedProducts} setNewProductId={this.props.setNewProductId} actionButtonIcon={faStar} actionClick={this.handleActionButton}/>
       {modal}
 
     </div>)
